+++ Src/Services/CaseService/Case.ServiceBus/ComplaintRegisterService/IComplaintRegisterServiceBus.cs
@@ -11,7 +11,7 @@ namespace Case.ServiceBus.ComplaintRegisterService
 
         Task<ResponseWithoutPaginationModel> GetPersonAgainstDetails(long ComplaintRegId);
         Task<ComplaintRegisterResponseModel> AddEditPersonAgainstDetails(PersonAgainstDetailsModel objModel, int UserId);
-        Task<ComplaintRegisterResponseModel> DeletePersonAgainstDetails(int PersonAgainstId, int UserId);
+        Task<ComplaintRegisterResponseModel> DeletePersonAgainstDetails(long PersonAgainstId, int UserId);
     }
